backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
ortools==9.8.3296
supabase==2.3.4
python-dotenv==1.0.1
pydantic==2.5.3
pydantic-settings==2.1.0
httpx>=0.24,<0.26
pytest==7.4.4
pytest-asyncio==0.23.3
